Update question3
Have been trying to push this for close to an hour but my local repository doesn't want to cooperate. Anyhow, thought it may be a good idea to add an associative array for each car. Each iteration of the for-loop adds an index of [car$i], so that later when the user is inputting we can refer to a specific lane (eg. car[car2]="|->                                       # Lane 2 #")