Create question4
The beginning of question 4, creates the savedPermission files with the permissions and names of files and directories in the working directory. Doesn't yet search the working directory recursively for subdirectories because I'm not clear if it has to. Also, creates 'savedPermissions0' as the first file rather than 'savedPermissions' due to how the name checking works.